+++ tests/testthat/test_objectinheritance.R
@@ -4,9 +4,9 @@ test_that('Check that distribution objects are properly inherited', {
   skip_if_not_installed('igraph')
   skip_if_not_installed('abind')
 
-  skip_if_not_installed("cmdstanr")
-  skip_if(condition = tryCatch(expr = cmdstanr::cmdstan_path(), error = function(e) return(TRUE)),
-          message = "No cmdstan path")
+  skip_if_not_installed("glmnet")
+  # skip_if(condition = tryCatch(expr = cmdstanr::cmdstan_path(), error = function(e) return(TRUE)),
+  #         message = "No cmdstan path")
 
   # Load packages
   suppressWarnings( requireNamespace("terra", quietly = TRUE) )
@@ -74,7 +74,7 @@ test_that('Check that distribution objects are properly inherited', {
   # Engine
   x |> engine_gdb(boosting_iterations = 500)
   expect_true(is.Waiver(x$engine))
-  x |> engine_stan()
+  x |> engine_glmnet()
   expect_true(is.Waiver(x$engine))
 
   # Priors